iwlwifi: mvm: BT Coex - add debugfs hook to set BT Tx priority
In order to debug the firmware, we need to be able to set
the BT priority of WiFi packets. This priority is set based
on the type of the packet (control frames, EAPOL etc...).
For debugging purposes, allow to override this priority by
a debugfs controlled value.
Enable this feature to be able to test it.

Change-Id: Ic9ff182085a855511cdd875cb63c58e1c95656b9
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/32367
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: ee40f888020f35fae0d549132bb0a45600305019