iwlwifi: mvm: rs: use correct max expected throughput figures
The selection of the max expected throughput for a column
didn't take into account the maximal allowed rate for the current
peer. This can cause unnecessary switches during the search cycle
to columns which have no chance of beating the current throughput.

Change-Id: Ib651f0b6e512744c3307d80ca741e050d223a8e7
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
While correcting the algorithm behavior this yields a very little TPT
improvement on my open air setup which is within the results normal variance
so hard to say if it really helps. Makes sense that it wouldn't
make much difference as it effects only switching which occurs during
the search cycle and most of the time we're not searching.
Pushing this for review.
This should be merged only after results from sanity performance setup verify
it doesn't degrade anything (I'll add this)
Reviewed-on: https://gerrit.rds.intel.com/34504
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34824
Tested-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 8fe211671446b01130222d3430860158afd2ad3a