iwlwifi: mvm: set the MAC_FILTER_IN_BEACON flag also for P2P
There doesn't seem to be a good reason for not enabling the
MAC_FILTER_IN_BEACON flag for P2P client, as we do for station.

This can prevent potential, hard-to-reproduce problems during
association.

Change-Id: I7c7f3d815cd3312447f352758ef3a99bf71224fb
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---

This was backported to Core5 where we don't have some of the
refactoring I did in the relevant functions.  The change done here is
the smallest change possible to achieve the same functional change
that the original patch does.

Change-Id: I7c7f3d815cd3312447f352758ef3a99bf71224fb
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37251
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/47301
Reviewed-by: Guy Cohen <guy.cohen@intel.com>
x-iwlwifi-stack-dev: 2cefc272465d224cfbe7acfef16e9ae1a7e8379c