mac80211: set new interfaces as idle upon init
Mark new interfaces as idle to allow operations that require that
interfaces are idle to take place. Interface types that are always
not idle (like AP interfaces) will be set as not idle when they are
assigned a channel context.

Change-Id: Ic3229a933ab3853d11c8c1ec2439ae6726dde5f1
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37567
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/37724

x-iwlwifi-stack-dev: 527ca4c343e5f76169ce8451a04ee314d97881ef