cfg80211: send events when devices are added/removed
We're currently sending NEW_WIPHY events for renames (which
is a bit odd, but now can't be changed), but also send them
for really new devices that register.

Also send DEL_WIPHY events when a device is removed, the
event ID for this was already reserved.

Change-Id: I8912ee643dbea93d1c8d040c187ca34e6a043e16
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37847
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Tested-by: IWL Jenkins
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37939

x-iwlwifi-stack-dev: 526bf7c14bade21bd35624052096febf0f3ce388