iwlwifi: mvm: don't use CTS to self
CTS to self is a partial protection only and it triggers
tricky flows in the firmware that break from time to time.
It doesn't have any real advantage besides a slight
improvement in throughput which is negligible compared to
the trouble it causes. Disable it altogether.

Change-Id: I792d03fa0e50a51a53596e5313e957545e32d6ff
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/41146
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/41604
Reviewed-by: Guy Cohen <guy.cohen@intel.com>

x-iwlwifi-stack-dev: c5e2013986dce0aa1824ae9fb1c45bdfd34e0105