[BUGFIX] nl80211: clear skb cb before passing to netlink
In testmode and vendor command reply/event SKBs we use the
skb cb data to store nl80211 parameters between allocation
and sending. This causes the code for CONFIG_NETLINK_MMAP
to get confused, because it takes ownership of the skb cb
data when the SKB is handed off to netlink, and it doesn't
explicitly clear it.

Clear the skb cb explicitly when we're done and before it
gets passed to netlink to avoid this issue.

Change-Id: I8763ee434b8d796f026dcf4b85a497ac79cdfd72
Cc: stable@vger.kernel.org [this goes way back]
Reported-by: Assaf Azulay <assaf.azulay@intel.com>
Reported-by: David Spinadel <david.spinadel@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
This bug surfaced in TCM events...
Reviewed-on: https://gerrit.rds.intel.com/41647
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Assaf Azulay <assaf.azulay@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: ArikX Nemtsov <arik@wizery.com>
Reviewed-by: David Spinadel <david.spinadel@intel.com>
Tested-by: David Spinadel <david.spinadel@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/41658

x-iwlwifi-stack-dev: d38d5bdd855519b1c7c7b8807886ad9914b0e9c5