iwlwifi: mvm: reset ucode_loaded flag when mac80211 stop is called
We rely on the value of the mvm->ucode_loaded flag to decide whether
or not we can perform certain operations (e.g. access to some debugfs
entries), so we need to reset it when the mac80211 stop operation is
called and the hardware is shutdown.

Change-Id: Ia7e0633a78b337b710e6905629b4d58b8465d3cf
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/42743
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43190
Reviewed-by: Guy Cohen <guy.cohen@intel.com>

x-iwlwifi-stack-dev: 17d8f3ce28caf1efe3fca8d21d6401a4a2bd2c84