iwlwifi: mvm: use the firmware to get the temperature during CT kill
Reading the temperature directly from the hardware, without the help
of the firmware, is a complex process and is not entirely the same for
different hardware.  Also, some NICs don't easily allow access to the
sensors when the firmware is not running, which would add even more
complexity to the code.

To reduce the code complexity and to avoid code duplication between
the firmware and the driver, boot the firmware briefly to read the
current temperature while in CT kill mode.

Change-Id: I304f6b25e5518924c6621aece763163cc302632f
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---
MWG100214268
Reviewed-on: https://gerrit.rds.intel.com/43528
Reviewed-by: EliadX Peller <eliad@wizery.com>
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43684
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 244dc763bce0a497d9985d341c69ef1b007ddb8c