iwlwifi: mvm: rs: fix logic in case of multiple TIDs
In case of traffic on multiple TIDs where one is aggregated
and the other is not RS would toggle between considering
traffic vs. the station as aggregated and not aggregated.
Instead consider the sta state as aggregated as long as
there's at least one TID aggregated.
This limitation is because the rates table is kept per
station and not per TID.

Change-Id: Ia16db1d63843eb3e56bfa000ba6eb8cc1f2425d6
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/44176
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/47791
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 1559652eca5e9aa4aa50a44502d435b1fe77dec1