iwlwifi: configure the LTR
The LTR is the handshake between the device and the root
complex about the latency allowed when the bus exits power
save. This configuration was missing and this led to high
latency in the link power up. The end user could experience
high latency in the network because of this.

Change-Id: Ic5f903ff57bb5ccc4a7aa6cab05030b2dbc1ecad
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/44858
Tested-by: IWL Jenkins
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/44938
Reviewed-by: Matt Chen <matt.chen@intel.com>
Tested-by: Matt Chen <matt.chen@intel.com>

x-iwlwifi-stack-dev: 310291c781d43af5278d7e325ae39660b975c489