Revert "iwlwifi: mvm: treat EAPOLs like mgmt frames wrt rate"
This reverts commit a5c159e1b8a6048b927325f61e8d5deee03b148b.

When the initial rate is _not_ chosen by RSSI conditions, the original
patch causes EAPOL frames to be sent with bad rates. This in turn
hinders connections to some APs.
The issue is most likely with using rates that are too low in p2p/11a
scenarios. Revert the patch until further debugging is done. This
make the behavior the same as before the RSSI-conditions patch.

Change-Id: I26d0c141e66a64dc176e172ad665362b08cc1037
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/45454
Tested-by: IWL Jenkins
Reviewed-by: Assaf Azulay <assaf.azulay@intel.com>
Tested-by: Assaf Azulay <assaf.azulay@intel.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/45786
Reviewed-on: https://gerrit.rds.intel.com/r/49614
x-iwlwifi-stack-dev: 8205fe2bc4a322b9342c8855492c999f5635bd52