mac80211: properly flush delayed scan work on interface removal
When an interface is deleted, an ongoing hardware scan is canceled and
the driver must abort the scan, at the very least reporting completion
while the interface is removed.

However, if it scheduled the work that might only run after everything
is said and done, which leads to cfg80211 warning that the scan isn't
reported as finished yet; this is no fault of the driver, it already
did, but mac80211 hasn't processed it.

To fix this situation, flush the delayed work when the interface being
removed is the one that was executing the scan.

Change-Id: Ib9bcd93fc5388510ec320065be28c2900bb08ca2
Cc: stable@vger.kernel.org
Reported-by: Sujith Manoharan <sujith@msujith.org>
Tested-by: Sujith Manoharan <sujith@msujith.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/46444
Reviewed-on: https://gerrit.rds.intel.com/r/47020
Tested-by: IWL Jenkins
Reviewed-by: Guy Cohen <guy.cohen@intel.com>
x-iwlwifi-stack-dev: 9676da2ddaab76fe84c9b10b9072abfa85869818