iwlwifi: mvm: rs: fix getting stuck in a test window
When Tx STBC is being used and RS switches to a search column
using the alternate antenna from the current one there is a
problem with using rs_rate_match to figure out which table
is the active and which one is the search one. The root cause
is because in STBC the antenna mask in the ucode rate is set
to ANT_AB and in this specific scenario it matches both the
active and search table (e.g. SISO_ANT_A and SISO_ANT_B).
This leads to tx stats being updated in the wrong table and later
on to getting stuck in a test window and not moving on to other
columns. If this happens during the initial search cycle we
never end it and therefore never enable aggregation which leads
naturally to severe througput degradation.
Fix it by deducing which table is which by knowing whether we're
in a search or not like it's being done in rs_rate_scale_perform

Change-Id: I3510df7c421f95e13a549db3942f63c0c4349a1c
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/46704
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/47040
x-iwlwifi-stack-dev: 4b71d5cccb86a0de909317a341d1ce1c458a7e4c