iwlwifi: mvm: pull SNAP header into skb->head
When we pre-populate the skb->head for the stack, we only pull
in the 802.11 header including crypto (assuming the packet isn't
short enough to be in there completely.) This is fine, but in
ieee80211_data_to_8023() we later unconditionally pull 8 more
bytes for the SNAP header and ethertype field (except for mesh
or 4-addr, where it's even more, but we don't care as much about
them).

Avoid the additional later pull by pulling in those 8 bytes here.

Change-Id: Ib37657ababa288089cc857ccf6b4f2d35bacb1b7
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/47144
Tested-by: IWL Jenkins
Reviewed-by: IdoX Yariv <ido@wizery.com>
Reviewed-on: https://gerrit.rds.intel.com/r/47766
x-iwlwifi-stack-dev: 76f0d845171ed19e24300eb1bd8b436d209f4d50