iwlwifi: mvm: rs: fix max rate allowed if no rate is allowed
In case the rate mask for one of the modulations was zero the
max rate idx for that modulation was set to 32 (BITS_PER_LONG).
This is bad as it would later lead to an out of bounds access
to the expected tpt table. In most cases there was no real effect
as the expected tpt was set to 0 and this led to avoiding the
modulation effectively.
Fix the out of bounds access and explicitly skip the modulation
in case there's no rate allowed in it.

Change-Id: Idf47680647327886c9eac62e84b7c569cced2554
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/48766
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57805
Reviewed-by: Guy Cohen <guy.cohen@intel.com>
x-iwlwifi-stack-dev: 63cbd9540445aa8626c09381a2c8866856b47f31