iwlwifi: sdio: Read HW_REV prior to FW load request
We must read the CSR_HW_REV in order to know which
FW to request.
The current step in the flow in which the CSR_HW_REV
is too late and we have already requested the FW from
the OS.
Reading a CSR requires a few init steps in order to have
the SDTM ready for commands.

Change-Id: Id89970de0a07df2c65a3e92dc11fadc98d441d63
Signed-off-by: Dor Shaish <dor.shaish@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/49121
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/49234
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 948323912e7a3fc191a06faf7e9c583016fc5e91