iwlwifi: mvm: tcm: convert load/global_load fields to enums
Clarify that 'load' and 'global_load' fields in the tcm result
are enums (representing the load level), and not an actual
load calculation.

Add iwl_mvm_tcm_load_percentage() helper function to calculate
the load percentage (which is used to determine the load level
enum to be stored). This function will be used in other places
later on.

Change-Id: I83788ccb148c1741ae8fb2fb541fec3a3742850f
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/50143
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/50634
x-iwlwifi-stack-dev: abaff9a6eac23065444c7155543121eb4e633626