mac80211: let flush() drop packets when possible [BUGFIX]
When roaming / suspending, it makes no sense to wait until
the transmit queues of the device are empty. In extreme
condition they can be starved (VO saturating the air), but
even in regular cases, it is pointless to delay the roaming
because the low level driver is trying to send packets to
an AP which is far away. We'd rather drop these packets and
let TCP retransmit if needed. This will allow to speed up
the roaming.

For suspend, the explanation is even more trivial.

Change-Id: I378600def9d373a1d3602bb7cd876a5106424de1
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
This is an attempt to solve
https://code.google.com/p/chrome-os-partner/issues/detail?id=34666
MWG100225076
Reviewed-on: https://gerrit.rds.intel.com/r/50405
Reviewed-by: Matt Chen <matt.chen@intel.com>
Tested-by: Matt Chen <matt.chen@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/50619
x-iwlwifi-stack-dev: 9d6b63282882a314516480871f44bf3a921b6fe4