iwlwifi: mvm: add the cause of the firmware dump in the dump
Now that the firmware dump can be triggered by events in
the code and not only the user or an firmware ASSERT, we
need a way to know why the firmware dump was triggered.
Add a section in the dump file for that.

Change-Id: Ic26fdae4db0324f969dbed507a670b66de823f4e
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/51869
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/52518
x-iwlwifi-stack-dev: cc6233ded25e54eb714d867cf8410e6dc2841304